+++ src/utils.ts
@@ -6,11 +6,6 @@ import { Mod } from '../typings/mc-curseforge-api/objects/Mod';
 import { ModFile } from '../typings/mc-curseforge-api/objects/Files';
 
 const config: BotConfig = Object.assign(configJson);
-const embedColors = [
-	0x404040,
-	0xFEBC11,
-	0xF26122,
-];
 
 type Release = 'Alpha' | 'Beta' | "Release";
 
@@ -31,21 +26,6 @@ export class Utils {
 	// cf getMod Wrapper function
 	static async getModById(id: number): Promise<Mod> {
 		return cf.getMod(id);
-	}
-
-	static createEmbed(title: string, description = ''): MessageEmbed {
-		const embed = new MessageEmbed();
-
-		embed.setTitle(title);
-
-		if (description !== '') {
-			embed.setDescription(description);
-		}
-
-		// Set a Random Embed Color
-		embed.setColor(embedColors[Math.ceil((Math.random() * 3))]);
-
-		return embed;
 	}
 
 	static async buildScheduleEmbed(guildId: Snowflake, client: Client): Promise<MessageEmbed> {
@@ -89,35 +69,6 @@ export class Utils {
 		}
 
 		return embed;
-	}
-
-	static buildModEmbed(mod: Mod, modFile: ModFile): MessageEmbed {
-		const modEmbed = new MessageEmbed();
-		const releaseTypeString: Release = this.getTypeStringFromId(modFile.release_type as unknown as number);
-		const splitUrl = modFile.download_url.split('/');
-		const fileName = splitUrl[splitUrl.length - 1];
-		let authorString = '';
-
-		for (let i = 0; i < mod.authors.length; i++) {
-			const author = mod.authors[i];
-			authorString += '[' + author.name + '](' + author.url + '), ';
-		}
-
-		modEmbed.type = 'rich';
-		modEmbed.setTitle('New ' + mod.name + ' ' + releaseTypeString + '!').setURL(mod.url);
-		modEmbed.setDescription(mod.summary);
-		modEmbed.addField('Authors', authorString);
-		modEmbed.setColor(releaseColors.get(releaseTypeString));
-		modEmbed.setThumbnail(mod.logo.url);
-		modEmbed.addField('Minecraft Versions', modFile.minecraft_versions.join(', '));
-		modEmbed.addField('New Mod Version', fileName);
-		modEmbed.addField('Type', releaseTypeString);
-		modEmbed.addField('Links', '[Download](' + modFile.download_url.replace(/ /g, '%20') + ')\n[CurseForge](' + mod.url + ')');
-		modEmbed.setTimestamp(modFile.timestamp);
-
-		console.log('Latest file: ' + fileName);
-
-		return modEmbed;
 	}
 
 	static async queryLatest(id: number): Promise<MessageEmbed> {
